Feat: Move screens from screens folder to components/pages to strictly follow Atomic Design convention
