+++ src/components/atoms/index.ts
@@ -1,0 +1,4 @@
+export {Button} from './Button';
+export {Icon} from './Icon';
+export {Input} from './Input';
+export {Logo} from './Logo';
